add mixin and update test
