+++ src/app/static/src/app/components/files/FileUpload.vue
@@ -22,9 +22,8 @@
 </template>
 
 <script lang="ts">
-    import Vue from "vue";
-    import {mapStatePropByName} from "../../utils";
-    import ResetConfirmation from "../ResetConfirmation.vue";
+    import ResetConfirmation from "../resetConfirmation/ResetConfirmation.vue";
+    import ResetConfirmationMixin from "../resetConfirmation/ResetConfirmationMixin"
 
     interface Methods {
         handleFileSelect: () => void
@@ -36,11 +35,6 @@
         showUploadConfirmation: boolean
     }
 
-    interface Computed {
-        editsRequireConfirmation: boolean
-        dataExplorationMode: boolean;
-    }
-
     interface Props {
         upload: (formData: FormData) => void,
         accept: string,
@@ -48,7 +42,7 @@
         uploading: boolean
     }
 
-    export default Vue.extend<Data, Methods, Computed, Props>({
+    export default ResetConfirmationMixin.extend<Data, Methods, unknown, Props>({
         props: {
             "upload": Function,
             "accept": String,
@@ -62,15 +56,6 @@
         },
         components: {
             ResetConfirmation
-        },
-        computed: {
-            dataExplorationMode: mapStatePropByName(null, "dataExplorationMode"),
-            editsRequireConfirmation() {
-                if (this.dataExplorationMode) {
-                    return false
-                }
-                return this.$store.getters["stepper/editsRequireConfirmation"]
-            }
         },
         methods: {
             handleFileSelect() {
