+++ src/app/static/src/tests/components/files/fileUploadWithEditConfirmation.test.ts
@@ -1,7 +1,7 @@
 import Vuex from "vuex";
 import {mount, Wrapper} from '@vue/test-utils';
 import FileUpload from "../../../app/components/files/FileUpload.vue";
-import ResetConfirmation from "../../../app/components/ResetConfirmation.vue";
+import ResetConfirmation from "../../../app/components/resetConfirmation/ResetConfirmation.vue";
 import {mockFile, mockRootState} from "../../mocks";
 import {emptyState, RootState} from "../../../app/root";
 import registerTranslations from "../../../app/store/translations/registerTranslations";
